maia-training-rl-main/requirements.txt
chess==1.11.2
dnspython==2.7.0
numpy==2.2.6
pymongo==4.13.2
python-dotenv==1.1.1